remvoe some usage of tendermint merkle
