+++ internal/proofs/create.go
@@ -5,7 +5,7 @@ import (
 	"sort"
 
 	ics23 "github.com/confio/ics23/go"
-	"github.com/tendermint/tendermint/crypto/merkle"
+	sdkmaps "github.com/cosmos/cosmos-sdk/internal/maps"
 )
 
 // TendermintSpec constrains the format from ics23-tendermint (crypto/merkle SimpleProof)
@@ -93,7 +93,7 @@ func createExistenceProof(data map[string][]byte, key []byte) (*ics23.ExistenceP
 		return nil, fmt.Errorf("cannot make existence proof if key is not in map")
 	}
 
-	_, ics23, _ := merkle.SimpleProofsFromMap(data)
+	_, ics23, _ := sdkmaps.SimpleProofsFromMap(data)
 	proof := ics23[string(key)]
 	if proof == nil {
 		return nil, fmt.Errorf("returned no proof for key")
