+++ store/rootmulti/store_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/store/iavl"
 	"github.com/cosmos/cosmos-sdk/store/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	sdkmaps "github.com/cosmos/cosmos-sdk/internal/maps"
 )
 
 func TestStoreType(t *testing.T) {
@@ -593,5 +594,5 @@ func hashStores(stores map[types.StoreKey]types.CommitKVStore) []byte {
 			},
 		}.GetHash()
 	}
-	return SimpleHashFromMap(m)
+	return sdkmaps.SimpleHashFromMap(m)
 }
